Move "everything" into Storage class
